+++ docs/changelog.rst
@@ -4,6 +4,14 @@
 Changelog
 *********
 
+0.3.1
+=====
+- train-pool for training parallelized at the level of trials rather than
+  computations
+- optional tqdm dependency (used by train-pool)
+- add separate joblib dependency (should be installed w/scikit-learn,
+  scikit.externals.joblib is deprecated)
+
 0.3.0
 =====